Close #9: Implement smooth car rearrangement
Now car takes two lanes while rearrangement
and moves smoothly along diagonal
(hypotenuse of an isosceles triangle equals dx`).
`dx` is a distanse the car will pass when changing frames.
Car can change lane
only if there are no cars both behind or forward it on new lane
(added `forward` argument in `isClose` function).
If 'forward` is `true` or not specified, the function chechs
if there are cars forward the current car and close to it.
If `forward` is `false`, the function checks
if there are cars behind the car and close to it.
If there are no close cars to the current car,
it can change lane to new (`car.newLane` gets number of new lane
and is not `undefined` before car reaches new lane).
Otherwise it stopes or moves on the current lane (`car.lane`).
When car reaches new lane (`car.laneChangeProgress` reaches `1`)
it takes only new lane.